==> 1-ETL.ipynb <==
## Descripcion

Este codigo consiste en realizar un ETL de los datos de la carpeta 'ignore':  
1-Extraccion: Importa 4 datasets CSV correspondientes a los titulos por plataforma y 8 datasets CSV correspondientes a las puntuaciones de titulos por usuario.  
2-Transformacion: se realizan 2 transformaciones  
    a) All: Busca unificar los titulos de las 4 plataformas para unificar las consultas en un dataset 'All'.  
    b) Score: Busca unificar las puntuaciones de titulos para consultas de score en un unico dataset 'Score'.   
3-Carga: Exporta los dataframes a los siguientes archivos:  
    a) All => all.parquet  
    b) Score => score.csv  
    c) Score => score.parquet  
    d) Cast => cast.csv  

## Extraccion

Importa 4 datasets correspondientes a los titulos por plataforma y 8 datasets correspondientes a las puntuaciones de titulos por usuario.

import pandas as pd
import numpy as np

# Titles
Amazon = pd.read_csv("ignore/amazon_prime_titles.csv")
Disney = pd.read_csv("ignore/disney_plus_titles.csv")
Hulu = pd.read_csv("ignore/hulu_titles.csv")
Netflix = pd.read_csv("ignore/netflix_titles.csv")

# Scores
a = pd.read_csv("ignore/ratings/1.csv")
b = pd.read_csv("ignore/ratings/2.csv")
c = pd.read_csv("ignore/ratings/3.csv")
d = pd.read_csv("ignore/ratings/4.csv")
e = pd.read_csv("ignore/ratings/5.csv")
f = pd.read_csv("ignore/ratings/6.csv")
g = pd.read_csv("ignore/ratings/7.csv")
h = pd.read_csv("ignore/ratings/8.csv")

# Obtiene informacion
print('Amazon')
print('Duplicates:',Amazon.duplicated().any())
print('Files, Columns:',Amazon.shape)
print(Amazon.isnull().sum())
Amazon.tail(3)


# Obtiene informacion
print('Disney')
print('Duplicates:',Disney.duplicated().any())
print('Files, Columns:',Disney.shape)
print(Disney.isnull().sum())
Disney.tail(3)

# Obtiene informacion
print('Hulu')
print('Duplicates:',Hulu.duplicated().any())
print('Files, Columns:',Hulu.shape)
print(Hulu.isnull().sum())
Hulu.tail(3)

# Obtiene informacion
print('Netflix')
print('Duplicates:',Netflix.duplicated().any())
print('Files, Columns:',Netflix.shape)
print(Netflix.isnull().sum())
Netflix.tail(3)

## Transformacion Titles

# Generar campo id: Cada id se compondrá de la primera letra del nombre de la plataforma, 
# seguido del show_id ya presente en los datasets (ejemplo para títulos de Amazon = as123)
id_A = 'a' + Amazon['show_id']
Amazon.insert(0, 'id', id_A)
Amazon.insert(1, 'platform', 'amazon')

id_D = 'd' + Disney['show_id']
Disney.insert(0, 'id', id_D)
Disney.insert(1, 'platform', 'disney')

id_H = 'h' + Hulu['show_id']
Hulu.insert(0, 'id', id_H)
Hulu.insert(1, 'platform', 'hulu')

id_N = 'n' + Netflix['show_id']
Netflix.insert(0, 'id', id_N)
Netflix.insert(1, 'platform', 'netflix')

Titles = pd.concat([Amazon,Disney,Hulu,Netflix], axis=0).reset_index()
print('Titles platforms')
print('Files, Columns:',Titles.shape)
print(Titles.isnull().sum())
Titles.head(3)


# Los valores nulos del campo rating deberán reemplazarse por el string “G” (corresponde al maturity rating: “general for all audiences”)
Titles['rating'].fillna('G',inplace=True)
print(Titles['rating'].isnull().sum())

# De haber fechas, deberán tener el formato AAAA-mm-dd
Titles['date_added'] = pd.to_datetime(Titles['date_added'])
Titles.tail(3)

# Los campos de texto deberán estar en minúsculas, sin excepciones
lowercase = lambda s:s.lower() if type(s) == str else s
Titles = Titles.applymap(lowercase)

# El campo duration debe convertirse en dos campos: duration_int y duration_type. 
# El primero será un integer y el segundo un string indicando la unidad de medición de duración: min (minutos) o season (temporadas)
Titles[['duration_int','duration_type']] = Titles['duration'].str.split(' ', expand=True)
Titles.drop(columns=['duration'], inplace=True)
Titles.nunique()          

# Organiza las columnas
Titles = Titles.reindex(columns=['id', 'platform', 'type', 'title', 'director', 'cast', 'country', 'date_added', 'release_year', 'duration_int', 'duration_type', 'listed_in', 'description'])
Titles.info()

# Cantidad de valores unicos
Titles.nunique()

# Convierte a integer los valores numericos
Titles['duration_int'] = Titles['duration_int'].fillna(0)
Titles['duration_int'] = Titles['duration_int'].astype(int)
Titles['release_year'] = Titles['release_year'].astype(int) 
Titles.info()

# VIsualiza que categorias de datos hay en duration_type
import collections
collections.Counter(Titles['duration_type'])

# Unifica seasons y season
Titles['duration_type'] = Titles['duration_type'].replace('seasons','season')
collections.Counter(Titles['duration_type'])

# Lista los encabezados
headers = list(Titles.columns)
print(headers)  

# Organiza y reduce las columnas
Titles = Titles.reindex(columns=['id', 'platform', 'type', 'title', 'director', 'cast', 'country', 'date_added', 'release_year', 'duration_int', 'duration_type', 'listed_in', 'description'])
Titles.info()

# Opcion de guardar avance parcial cambiando switch a True (solo con carpeta ignore)
switch = False
if switch:
    Titles.to_csv('ignore/all_titles.csv', index=False)

## Transformacion Score

# Unifica los dataset de rating en Score
Score = pd.concat([a,b,c,d,e,f,g,h], axis=0)
Score = Score.reset_index(drop=True)
print('Score')
print('Files, Columns:',Score.shape)
print(Score.isnull().sum())
Score.tail(3)

# Reemplaza timestamp por date, agrega la columna platform, renombra rating a score
Score = Score.rename(columns={'rating':'score'})                        # Renombra rating con score
Score['date'] = pd.to_datetime(Score['timestamp'], unit='s').dt.date    # Agrega columna date
Score.drop(columns='timestamp', inplace=True)                           # Elimina timestamp
Score['platform'] = Score['movieId'].str[0]                             # Agrega la columna platform
# Score = Score[['platform','score','year','movieId']]
Score.tail(3)

# Completa el nombre de las plataformas mediante una funcion
def get_initial(movieId:str):
    if movieId[0] == 'a':
        platform = 'amazon'
    elif movieId[0] == 'd':
        platform = 'disney'
    elif movieId[0] == 'h':
        platform = 'hulu'
    elif movieId[0] == 'n':
        platform = 'netflix'
    else: platform = '' 
    
    return platform


# Aplica la función a la columna 'platform' para reemplazar las iniciales con el nombre de la plataforma
Score['platform'] = Score['platform'].apply(get_initial)
Score.tail(2)

# Reordena las columnas y elimina date
Score = Score[['userId','movieId','score','platform']]
Score.tail(2)

# Define una función de agregación que calcula los promedios del score y consulta valores estadisticos
S = Score.groupby('movieId').agg(pd.Series({'score': 'mean'}))
S.describe()

Titles.tail(2)

first_release = Titles['release_year'].min()
max_duration = Titles['duration_int'].max()
users = Score['userId'].max()
print(first_release,max_duration,users)

## Transformacion All

# Filtra las columnas userId, movieId y score del dataset Sore y las columnas id y title del dataset Title.
Score = Score[['userId','movieId','score','platform']]
Titles = Titles.rename(columns={'id':'movieId'}) # Cambia el nombre para que haya coincidencia
title = Titles[['movieId','title']]     # Guarda solo los titulos con su id
Titles.tail(2)

All = pd.merge(Score,Titles,how='left',on='movieId')
scores_amount = len(All)-1
All.tail(2)

## Carga

# Opcion de cargar avance parcial cambiando switch a True (solo con carpeta ignore)
switch = False
if switch:
    all = pd.read_csv('ignore/all_titles.csv')
    Score = pd.read_parquet('processed_data/score.parquet')

# S = Score.groupby('movieId').agg(pd.Series({'score': 'mean'}))
# S.tail(2)

all = all.rename(columns={'id':'movieId'}) 

all = pd.merge(S,all,how='left',on='movieId')

all = all[['score','platform', 'title', 'cast', 'release_year', 'duration_int', 'duration_type']]
all.to_csv('app/data/all.csv')


# Titles['platform'] = Titles['platform'].astype(str)
# Titles['type'] = Titles['type'].astype(str)
# Titles['title'] = Titles['title'].astype(str)
# Titles['director'] = Titles['director'].astype(str)
# Titles['cast'] = Titles['cast'].astype(str)
# Titles['country'] = Titles['country'].astype(str)
# Titles['type'] = Titles['type'].astype(str)
# Titles['title'] = Titles['title'].astype(str)
# Titles['director'] = Titles['director'].astype(str)
Titles['release_year'] = Titles['release_year'].astype(np.int16)
Titles['duration_int'] = Titles['duration_int'].astype(np.int16)
# Titles['listed_in'] = Titles['listed_in'].astype(str)
# Titles['description'] = Titles['description'].astype(str)

Score['userId'] = Score['userId'].astype(np.int32)
Score['score'] = Score['score'].astype(np.int8)
# Score['platform'] = Score['platform'].astype(str)

All.info()

Score.info()

all.to_csv('data/all.csv')

Titles.to_parquet('processed_data/titles.parquet')

title.to_parquet('processed_data/title.parquet')

# Opcion de exportar csv de gran tamanio cambiando switch a True (solo con carpeta ignore)
switch = False
if switch:
    Score.to_csv('ignore/score.csv', index=False) 
# Da un archivo de 300Mb

# S.to_csv('data/score_mean.csv', index=False)

# Convierte score.csv en score.parquet para reducir el tamanio del archivo
# Score = pd.read_csv('ignore/score.csv')
Score.to_parquet('processed_data/score.parquet')

S.to_parquet('processed_data/S.parquet')

All.to_parquet('processed_data/all.parquet')
