==> 2-Queries.ipynb <==
## Descripcion

## Load Data

import pandas as pd

# Importa titles
All = pd.read_parquet('processed_data/titles.parquet')
# All = pd.read_csv("data/all.csv")

Score = pd.read_parquet('processed_data/score.parquet')
# Score_CSV = pd.read_csv('ignore/score.csv')

Score.tail(3)

# Obtiene informacion
print('All platforms')
print('Duplicates:',All.duplicated().any())
print('Files, Columns:',All.shape)
print(All.isnull().sum())
All.tail(2)

## Queries

# Variables que tomarán el valor de los parámetros de filtro de las queries.
platform = 'hulu'
type = 'movie'
scored = 2
year = 2019
duration_type = 'min'

### Query 1

# Prueba los filtros antes de hacer la funcion Query 1
# Película con mayor duración con filtros opcionales de AÑO, PLATAFORMA Y TIPO DE DURACIÓN. 
# (la función debe llamarse get_max_duration(year, platform, duration_type))
# Q1 = All[(All['platform'] == platform) & (All['release_year'] == year) & (All['duration_type'] == duration_type)].sort_values(by = 'duration_int' , ascending=False)
# Q1['title'].values[0]

# Funcion Query 1
# Devuelve: Película con mayor duración con filtros opcionales de AÑO, PLATAFORMA Y TIPO DE DURACIÓN. 
def get_max_duration(year:int, platform:str, duration_type:str): 
    Q1 = All[(All['platform'] == platform) & (All['release_year'] == year) & (All['duration_type'] == duration_type)].sort_values(by = 'duration_int' , ascending=False)
    if len(Q1) > 0:
        result = Q1['title'].values[0] 
        return result
    else:
        return ("Not titles found with those parameters.")
    
get_max_duration(year, platform, duration_type)

### Query 2

# Prueba los filtros antes de hacer la funcion Query 2
# Cantidad de películas por plataforma con un puntaje mayor a XX en determinado año 
# (la función debe llamarse get_score_count(platform, scored, year))
# Score[(Score['platform'] == platform) & (Score['score'] >= scored) & (All['release_year'] == year)].shape[0]

# Funcion Query 2
def get_score_count(platform, scored, year):
    Q2 = Score[(Score['platform'] == platform) & (Score['score'] >= scored) & (All['release_year'] == year)].shape[0]
    return Q2
      
get_score_count(platform, scored, year)  

### Query 3

# Funcion Query 3
# Cantidad de películas por plataforma con filtro de PLATAFORMA. (La función debe llamarse get_count_platform(platform))
def get_count_platform(platform:str, All:pd.DataFrame):
    All = All.loc[All['platform'] == platform]
    movies_platform = len(All.index)
    return movies_platform

get_count_platform(platform, All)


### Query 4

# Actor que más se repite según plataforma y año. (La función debe llamarse get_actor(platform, year))
Not_found = {'actor': 'Not found', 'appearances': 0}
Q4_1 = All[(All['platform'] == platform) & (All['release_year'] == year)]   # Filtra por plataforma y anio
if len(Q4_1) == 0:                              # Si no hay valores para esos filtros: No encontrado
    print(Not_found)
else:    
    Q4_2 = Q4_1.assign(actor=Q4_1.cast.str.split(',')).explode('cast')  # Divide 'cast' por comas
    mask = Q4_2['actor'].isna()                 # Filtra los valores nulos
    Q4_3 = Q4_2.loc[mask, 'actor']
    if len(Q4_3) == 0:                          # Si no hay valores luego de filtrar nulos: No encontrado
        print(Not_found)
    else:
        Q4_4 = Q4_3.cast.value_counts()         # Crea una lista con actores y apariciones.
        max_actor = Q4_4.index[0]               # Obtiene el actor con más apariciones
        max_count = int(Q4_4.iloc[0])           # Obtiene el número de apariciones
        Q5 = dict({'actor': max_actor, 'appearances': max_count})
        print(Q5)
    


mask = Q4_2['actor'].isna()
Q4_3 = Q4_2.loc[mask, 'actor']
len(Q4_3)

# Funcion Query 4
def get_actor(platform:str, year:int):
    All = pd.read_csv("data/all.csv")
    Q4_1 = All[(All['platform'] == platform) & (All['release_year'] == year)]   # Filtra por plataforma y anio
    if len(Q4_1) > 0: 
        Q4_2 = Q4_1.assign(actor=Q4_1.cast.str.split(',')).explode('cast')  # Divide 'cast' por comas
        Q4_3 = Q4_2.cast.value_counts()         # Crea una lista con actores y apariciones.
        max_actor = Q4_3.index[0]               # Obtiene el actor con más apariciones
        max_count = int(Q4_3.iloc[0])           # Obtiene el número de apariciones
        if max_actor == 'nan':                  # Si los valores nulos son los mas frecuentes
            max_actor = Q4_3.index[1]           # Obtiene el segundo actor con más apariciones
            max_count = int(Q4_3.iloc[1])       # Obtiene el número de apariciones correspondiente
        Q4 = dict({'actor': max_actor, 'appearances': max_count})
        return Q4
    else:
        return ({'actor': '', 'appearances': 0})
    
get_actor(platform, year)

# All = pd.read_csv("data/all.csv")
# All = All.loc[All['release_year'] == year]
# All = All.loc[All['platform'] == platform]
# All = All.loc[All['duration_type'] == duration_type]
# All['duration_int'].max()


# All.to_csv('data/all.csv', index=False)
